Fixed block malleability test and snapshot test
